requirements.txt
Flask
openai
pandas
langchain
langchain-openai
python-dotenv
